requirements.txt
Django==2.1
django-filter==2.0.0
djangorestframework==3.8.2
Markdown==2.6.11